migrated dimension related queries to dimensions.go file, and tests to dimensions_test.go. Implemented unhappy paths for all the calls.
