+++ cantabular/dimensions_test.go
@@ -1,0 +1,371 @@
+package cantabular_test
+
+import (
+	"context"
+	"io"
+	"net/http"
+	"testing"
+
+	"github.com/ONSdigital/dp-api-clients-go/v2/cantabular"
+	dphttp "github.com/ONSdigital/dp-net/http"
+	. "github.com/smartystreets/goconvey/convey"
+)
+
+func TestGetDimensionsHappy(t *testing.T) {
+	Convey("Given a correct getDimensions response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		mockHttpClient, cantabularClient := newMockedClient(mockRespBodyGetDimensions, http.StatusOK)
+
+		Convey("When GetDimensions is called", func() {
+			resp, err := cantabularClient.GetDimensions(testCtx, "Teaching-Dataset")
+
+			Convey("Then no error should be returned", func() {
+				So(err, ShouldBeNil)
+			})
+
+			Convey("And the expected query is posted to cantabular api-ext", func() {
+				So(mockHttpClient.PostCalls(), ShouldHaveLength, 1)
+				So(mockHttpClient.PostCalls()[0].URL, ShouldEqual, "cantabular.ext.host/graphql")
+				validateQuery(
+					mockHttpClient.PostCalls()[0].Body,
+					cantabular.QueryDimensions,
+					cantabular.QueryData{
+						Dataset: "Teaching-Dataset",
+					},
+				)
+			})
+
+			Convey("And the expected response is returned", func() {
+				So(*resp, ShouldResemble, expectedDimensions)
+			})
+		})
+	})
+}
+
+func TestGetDimensionsUnhappy(t *testing.T) {
+	Convey("Given a no-dataset graphql error response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespBodyNoDataset, http.StatusOK)
+
+		Convey("When GetDimensions is called", func() {
+			resp, err := cantabularClient.GetDimensions(testCtx, "InexistentDataset")
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoDatasetErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a 500 HTTP Status response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespInternalServerErr, http.StatusInternalServerError)
+
+		Convey("When GetDimensions is called", func() {
+			resp, err := cantabularClient.GetDimensions(testCtx, "Teaching-Dataset")
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedInternalServeError)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+}
+
+func TestGetGeographyDimensionsHappy(t *testing.T) {
+	Convey("Given a correct getGeographyDimensions response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		mockHttpClient, cantabularClient := newMockedClient(mockRespBodyGetGeographyDimensions, http.StatusOK)
+
+		Convey("When GetGeographyDimensions is called", func() {
+			resp, err := cantabularClient.GetGeographyDimensions(testCtx, "Teaching-Dataset")
+
+			Convey("Then no error should be returned", func() {
+				So(err, ShouldBeNil)
+			})
+
+			Convey("And the expected query is posted to cantabular api-ext", func() {
+				So(mockHttpClient.PostCalls(), ShouldHaveLength, 1)
+				So(mockHttpClient.PostCalls()[0].URL, ShouldEqual, "cantabular.ext.host/graphql")
+				validateQuery(
+					mockHttpClient.PostCalls()[0].Body,
+					cantabular.QueryGeographyDimensions,
+					cantabular.QueryData{
+						Dataset: "Teaching-Dataset",
+					},
+				)
+			})
+
+			Convey("And the expected response is returned", func() {
+				So(*resp, ShouldResemble, expectedGeographyDimensions)
+			})
+		})
+	})
+}
+
+func TestGetGeographyDimensionsUnhappy(t *testing.T) {
+	Convey("Given a no-dataset graphql error response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespBodyNoDataset, http.StatusOK)
+
+		Convey("When GetGeographyDimensions is called", func() {
+			resp, err := cantabularClient.GetGeographyDimensions(testCtx, "InexistentDataset")
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoDatasetErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a 500 HTTP Status response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespInternalServerErr, http.StatusInternalServerError)
+
+		Convey("When GetGeographyDimensions is called", func() {
+			resp, err := cantabularClient.GetGeographyDimensions(testCtx, "Teaching-Dataset")
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedInternalServeError)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+}
+
+func TestGetDimensionsByNameHappy(t *testing.T) {
+	Convey("Given a correct getDimensionsByName response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		mockHttpClient, cantabularClient := newMockedClient(mockRespBodyGetDimensionsByName, http.StatusOK)
+
+		Convey("When GetDimensionsByName is called", func() {
+			resp, err := cantabularClient.GetDimensionsByName(testCtx, cantabular.GetDimensionsByNameRequest{
+				Dataset:        "Teaching-Dataset",
+				DimensionNames: []string{"Age", "Region"},
+			})
+
+			Convey("Then no error should be returned", func() {
+				So(err, ShouldBeNil)
+			})
+
+			Convey("And the expected query is posted to cantabular api-ext", func() {
+				So(mockHttpClient.PostCalls(), ShouldHaveLength, 1)
+				So(mockHttpClient.PostCalls()[0].URL, ShouldEqual, "cantabular.ext.host/graphql")
+				validateQuery(
+					mockHttpClient.PostCalls()[0].Body,
+					cantabular.QueryDimensionsByName,
+					cantabular.QueryData{
+						Dataset:   "Teaching-Dataset",
+						Variables: []string{"Age", "Region"},
+					},
+				)
+			})
+
+			Convey("And the expected response is returned", func() {
+				So(*resp, ShouldResemble, expectedDimensionsByName)
+			})
+		})
+	})
+}
+
+func TestGetDimensionsByNameUnhappy(t *testing.T) {
+	testCtx := context.Background()
+
+	Convey("Given a no-dataset graphql error response from the /graphql endpoint", t, func() {
+		_, cantabularClient := newMockedClient(mockRespBodyNoDataset, http.StatusOK)
+
+		Convey("When the GetDimensionsByName method is called", func() {
+			req := cantabular.GetDimensionsByNameRequest{
+				Dataset:        "InexistentDataset",
+				DimensionNames: []string{"Country", "Age", "Occupation"},
+			}
+			resp, err := cantabularClient.GetDimensionsByName(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoDatasetErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a no-variable graphql error response from the /graphql endpoint", t, func() {
+		_, cantabularClient := newMockedClient(mockRespBodyNoVariable, http.StatusOK)
+
+		Convey("When the GetDimensionOptions method is called", func() {
+			req := cantabular.GetDimensionsByNameRequest{
+				Dataset:        "Teaching-Dataset",
+				DimensionNames: []string{"Country", "Age", "inexistentVariable"},
+			}
+			resp, err := cantabularClient.GetDimensionsByName(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoVariableErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a 500 HTTP Status response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespInternalServerErr, http.StatusInternalServerError)
+
+		Convey("When GetDimensionsByName is called", func() {
+			req := cantabular.GetDimensionsByNameRequest{
+				Dataset:        "Teaching-Dataset",
+				DimensionNames: []string{"Age", "Region"},
+			}
+			resp, err := cantabularClient.GetDimensionsByName(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedInternalServeError)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+}
+
+func TestGetDimensionOptionsHappy(t *testing.T) {
+	Convey("Given a correct getDimensionOptions response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		mockHttpClient, cantabularClient := newMockedClient(mockRespBodyGetDimensionOptions, http.StatusOK)
+
+		Convey("When GetDimensionOptions is called", func() {
+			req := cantabular.GetDimensionOptionsRequest{
+				Dataset:   "Teaching-Dataset",
+				Variables: []string{"Country", "Age", "Occupation"},
+			}
+			resp, err := cantabularClient.GetDimensionOptions(testCtx, req)
+
+			Convey("Then no error should be returned", func() {
+				So(err, ShouldBeNil)
+			})
+
+			Convey("And the expected query is posted to cantabular api-ext", func() {
+				So(mockHttpClient.PostCalls(), ShouldHaveLength, 1)
+				So(mockHttpClient.PostCalls()[0].URL, ShouldEqual, "cantabular.ext.host/graphql")
+				validateQuery(
+					mockHttpClient.PostCalls()[0].Body,
+					cantabular.QueryDimensionOptions,
+					cantabular.QueryData{
+						Dataset:   "Teaching-Dataset",
+						Variables: []string{"Country", "Age", "Occupation"},
+					},
+				)
+			})
+
+			Convey("And the expected response is returned", func() {
+				So(*resp, ShouldResemble, expectedDimensionOptions)
+			})
+		})
+	})
+}
+
+func TestGetDimensionOptionsUnhappy(t *testing.T) {
+	testCtx := context.Background()
+
+	Convey("Given a no-dataset graphql error response from the /graphql endpoint", t, func() {
+		_, cantabularClient := newMockedClient(mockRespBodyNoDataset, http.StatusOK)
+
+		Convey("When the GetDimensionOptions method is called", func() {
+			req := cantabular.GetDimensionOptionsRequest{
+				Dataset:   "InexistentDataset",
+				Variables: []string{"Country", "Age", "Occupation"},
+			}
+			resp, err := cantabularClient.GetDimensionOptions(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoDatasetErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a no-variable graphql error response from the /graphql endpoint", t, func() {
+		_, cantabularClient := newMockedClient(mockRespBodyNoVariable, http.StatusOK)
+
+		Convey("When the GetDimensionOptions method is called", func() {
+			req := cantabular.GetDimensionOptionsRequest{
+				Dataset:   "Teaching-Dataset",
+				Variables: []string{"Country", "Age", "inexistentVariable"},
+			}
+			resp, err := cantabularClient.GetDimensionOptions(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedNoVariableErr)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+
+	Convey("Given a 500 HTTP Status response from the /graphql endpoint", t, func() {
+		testCtx := context.Background()
+		_, cantabularClient := newMockedClient(mockRespInternalServerErr, http.StatusInternalServerError)
+
+		Convey("When GetDimensionOptions is called", func() {
+			req := cantabular.GetDimensionOptionsRequest{
+				Dataset:   "Teaching-Dataset",
+				Variables: []string{"Country", "Age", "Occupation"},
+			}
+			resp, err := cantabularClient.GetDimensionOptions(testCtx, req)
+
+			Convey("Then the expected error is returned", func() {
+				So(err, ShouldResemble, expectedInternalServeError)
+			})
+
+			Convey("And no response is returned", func() {
+				So(resp, ShouldBeNil)
+			})
+		})
+	})
+}
+
+// newMockedClient creates a new cantabular client with a mocked response for post requests,
+// according to the provided response string and status code.
+func newMockedClient(response string, statusCode int) (*dphttp.ClienterMock, *cantabular.Client) {
+	mockHttpClient := &dphttp.ClienterMock{
+		PostFunc: func(ctx context.Context, url string, contentType string, body io.Reader) (*http.Response, error) {
+			return Response(
+				[]byte(response),
+				statusCode,
+			), nil
+		},
+	}
+
+	cantabularClient := cantabular.NewClient(
+		cantabular.Config{
+			Host:       "cantabular.host",
+			ExtApiHost: "cantabular.ext.host",
+		},
+		mockHttpClient,
+		nil,
+	)
+
+	return mockHttpClient, cantabularClient
+}
